mcp-server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.1
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
celery==5.3.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiohttp==3.9.0
prometheus-client==0.19.0
python-dotenv==1.0.0
websockets==12.0
tenacity==8.2.3
structlog==23.2.0 